requirements.txt
confluent-kafka>=2.3.0
PyYAML>=6.0.1

